Move storage helpers to fix cypress test
